[feat]: party room admin can now remove other members from the party. fix on partymember model. fix on delete partyrooms on controller and api routes

+++ app/Http/Controllers/PartyRoomController.php
@@ -95,28 +95,45 @@ class PartyRoomController extends Controller
         }
     }
 
-    public function deletePartyMember(Request $request, $id)
+    public function deletePartyMember(Request $request)
     {
         try {
-           $userId = auth()->user()->id;
+            $userId = auth()->user()->id;
 
-           $partyMember = PartyMember::query()
-           ->where("id", $id)
-           ->where("user_id", $userId)
-           ->first();
+            $partyRoomId = $request->input('partyRoomId');
+            $partyMemberId = $request->input('partyMemberId');
 
-           $partyMember->delete();
+            $partyMember = PartyMember::with('party_rooms')
+                ->where('id', $partyMemberId)
+                ->whereHas('party_rooms', function ($query) use ($partyRoomId, $userId) {
+                    $query->where('id', $partyRoomId)
+                        ->where('admin_id', $userId);
+                })
+                ->first();
 
-           return response()->json(
-            [
-                "success" => true,
-                "message" => "PartyMember deleted successfully",
-                "data" => $partyMember
-            ],
-            Response::HTTP_OK
-        );
-} catch (\Throwable $th) {
+            if ($partyMember) {
+                $partyMember->delete();
+
+                return response()->json(
+                    [
+                        "success" => true,
+                        "message" => "PartyMember deleted successfully",
+                        "data" => $partyMember
+                    ],
+                    Response::HTTP_OK
+                );
+            } else {
+                return response()->json(
+                    [
+                        "success" => false,
+                        "message" => "Unauthorized to delete PartyMember from this PartyRoom",
+                    ],
+                    Response::HTTP_UNAUTHORIZED
+                );
+            }
+        } catch (\Throwable $th) {
             Log::error($th->getMessage());
+            Log::error($th->getTraceAsString()); 
 
             return response()->json(
                 [
@@ -127,6 +144,37 @@ class PartyRoomController extends Controller
             );
         }
     }
-}
-           
 
+    //     public function deletePartyMember(Request $request, $id)
+    //     {
+    //         try {
+    //            $userId = auth()->user()->id;
+
+    //            $partyMember = PartyMember::query()
+    //            ->where("id", $id)
+    //            ->where("user_id", $userId)
+    //            ->first();
+
+    //            $partyMember->delete();
+
+    //            return response()->json(
+    //             [
+    //                 "success" => true,
+    //                 "message" => "PartyMember deleted successfully",
+    //                 "data" => $partyMember
+    //             ],
+    //             Response::HTTP_OK
+    //         );
+    // } catch (\Throwable $th) {
+    //             Log::error($th->getMessage());
+
+    //             return response()->json(
+    //                 [
+    //                     "success" => false,
+    //                     "message" => "Error deleting PartyMember",
+    //                 ],
+    //                 Response::HTTP_INTERNAL_SERVER_ERROR
+    //             );
+    //         }
+    //     }
+}
